Squashed commit of the following:
commit 92a2a3adfe5cc62c26b8f2683e522e773fa4f49d
Merge: 1c4ee52 909de82
Author: Kalyan Kumar <kalyanjkk@gmail.com>
Date:   Mon Feb 24 14:37:53 2014 +0530

    Merge branch 'develop' of https://github.com/HSA-Libraries/Bolt into develop

commit 1c4ee52c310861ddd298012c0a47268953360e34
Author: Kalyan Kumar <kalyanjkk@gmail.com>
Date:   Mon Feb 24 14:37:14 2014 +0530

    fixed inner product test case

commit 909de82eaacfc7f2f69b48b24f126228ea4596fe
Merge: 1ea7e80 82fc8d5
Author: avinashcpandey <avinashcpandey@gmail.com>
Date:   Fri Feb 21 14:53:01 2014 +0530

    Merge pull request #156 from avinashcpandey/develop

    Compilation fix for EPR 394905

commit 82fc8d549aa1395840c26550b3a413229fd420e2
Author: avinashcpandey <avinashcpandey@gmail.com>
Date:   Fri Feb 21 01:21:14 2014 -0800

    Compilation fix for EPR 394905

commit 1ea7e802498c7bcd038c66f80403c2483730931b
Merge: df23c65 0608175
Author: avinashcpandey <avinashcpandey@gmail.com>
Date:   Thu Feb 20 20:43:10 2014 +0530

    Merge pull request #155 from avinashcpandey/develop

    Some comilation fixes

commit 0608175fc1ca0c691a4bb23a7966b2d7c49a1200
Author: avinashcpandey <avinashcpandey@gmail.com>
Date:   Thu Feb 20 07:10:51 2014 -0800

    Some comilation fixes

commit df23c65f09845bf43f602a3e3ab2c2e5ff6b1575
Author: Kalyan Kumar <kalyanjkk@gmail.com>
Date:   Wed Feb 19 14:25:17 2014 +0530

    fixed transform reduce

commit f370c0f8420dfbb2e0d60ebf937e807424387ef3
Author: Kalyan Kumar <kalyanjkk@gmail.com>
Date:   Mon Feb 17 17:27:32 2014 +0530

    Fixed compilation errors

commit 7844340921b8d03424317dcd64e15fba9f7ab5fe
Author: Kalyan Kumar <kalyanjkk@gmail.com>
Date:   Mon Feb 17 16:53:31 2014 +0530

    fixed compilation errors for 32 bit vs2013

commit d0401c8fb89ec053eed815a090f9cc3a51dfc798
Author: Kalyan Kumar <kalyanjkk@gmail.com>
Date:   Mon Feb 17 15:44:10 2014 +0530

    Fixed all compilation errors

commit b1fd70b07b4801978a8c093cb9e89233112ad330
Merge: 3dc7387 59d2107
Author: avinashcpandey <avinashcpandey@gmail.com>
Date:   Mon Feb 17 14:40:13 2014 +0530

    Merge pull request #154 from avinashcpandey/develop

    AMP MergeSortByKey fix for 65535 tile limit issue + EPR 394528  394534

commit 59d2107b0e6033b3240461c265453229e8d9fd4b
Author: avinashcpandey <avinashcpandey@gmail.com>
Date:   Mon Feb 17 01:08:19 2014 -0800

    AMP MergeSortByKey fix for 65535 tile limit issue + EPR 394528  394534

+++ test/amp/ReduceTest/ReduceTest.cpp
@@ -504,7 +504,7 @@ public:
         size_t size = GetParam( );
 
         std::generate(stdInput, stdInput + size, generateRandom<int>);
-        for (int i = 0; i<size; i++)
+		for (size_t i = 0; i<size; i++)
         {
             boltInput[i] = stdInput[i];
             boltOutput[i] = stdInput[i];
@@ -542,7 +542,7 @@ public:
         size_t size = GetParam( );
 
         std::generate(stdInput, stdInput + size, generateRandom<float>);
-        for (int i = 0; i<size; i++)
+		for (size_t i = 0; i<size; i++)
         {
             boltInput[i] = stdInput[i];
             boltOutput[i] = stdInput[i];
@@ -2068,9 +2068,9 @@ void testTBBDevicevector()
  //   bolt::amp::device_vector<int> tbbInput(aSize, 0);
     bolt::amp::device_vector<int> tbbInput(aSize);
 
-    for(int i=0; i<aSize; i++) {
-        stdInput[i] = i;
-        tbbInput[i] = i;
+	for (size_t i = 0; i<aSize; i++) {
+        stdInput[i] = (int)i;
+        tbbInput[i] = (int)i;
     };
 
     int hSum = std::accumulate(stdInput.begin(), stdInput.end(), 0);
